# src/swagger/dc-con-api.yml
swagger: "2.0"
info:
  title: "Domain Controller Connector"
  version: "1.0.0"
  description: "Domain controller connector interface."
  contact:
    name: "Christian Bremer"
    email: "bremersee@googlemail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "api.bremersee.org"
basePath: "/dc-con"
...
schemes:
  - "http"
  - "https"
paths:
...
  /api/groups:
    get:
      tags:
        - "domain-controller-connector"
      summary: "Get all domain groups."
      description: "Get all domain groups."
      operationId: "getGroups"
      produces:
        - "application/json"
      responses:
        200:
          description: "A list with domain group items."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/DomainGroupItem"
        500:
          description: "Fatal server error."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
    post:
      tags:
        - "domain-controller-connector"
      summary: "Add a domain group."
      operationId: "addGroup"
      parameters:
        - name: "group"
          description: "The domain group to add."
          in: "body"
          schema:
            $ref: "#/definitions/DomainGroup"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "The added domain group."
          schema:
            $ref: "#/definitions/DomainGroup"
        400:
          description: "Bad request."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        406:
          description: "Already exists"
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        500:
          description: "Fatal server error."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
  /api/groups/{groupName}:
    get:
      tags:
      - "domain-controller-connector"
      summary: "Get a domain group by name."
      operationId: "getGroupByName"
      parameters:
        - name: "groupName"
          description: "The domain group name."
          type: "string"
          required: true
          in: "path"
      produces:
        - "application/json"
      responses:
        200:
          description: "The domain group with the specified name."
          schema:
            $ref: "#/definitions/DomainGroup"
        400:
          description: "Bad request."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        404:
          description: "Not found"
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        500:
          description: "Fatal server error."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
    delete:
      tags:
        - "domain-controller-connector"
      summary: "Delete a domain group."
      operationId: "deleteGroup"
      parameters:
        - name: "groupName"
          description: "The domain group name."
          type: "string"
          required: true
          in: "path"
      produces:
        - "application/json"
      responses:
        200:
          description: "Group does not exist or was successfully deleted."
        400:
          description: "Bad request."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        500:
          description: "Fatal server error."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
  /api/groups/{groupName}/members:
    put:
      tags:
        - "domain-controller-connector"
      summary: "Update domain group members."
      operationId: "updateGroupMembers"
      parameters:
        - name: "groupName"
          description: "The domain group name."
          type: "string"
          required: true
          in: "path"
        - name: "members"
          description: "The domain group members."
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/Names"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "The updated domain group."
          schema:
            $ref: "#/definitions/DomainGroup"
        400:
          description: "Bad request."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        404:
          description: "Group not found."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        500:
          description: "Fatal server error."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
  /api/users:
    post:
      tags:
        - "domain-controller-connector"
      summary: "Add a domain user."
      operationId: "addUser"
      parameters:
        - name: "domainUser"
          description: "The domain user to add."
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/DomainUserAddRequest"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "The added domain user."
          schema:
            $ref: "#/definitions/DomainUser"
        400:
          description: "Bad request."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        500:
          description: "Fatal server error."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
  /api/users/{userName}:
    get:
      tags:
        - "domain-controller-connector"
      summary: "Get a domain user by name."
      operationId: "getUser"
      parameters:
        - name: "userName"
          description: "The user name of the domain user."
          type: "string"
          required: true
          in: "path"
      produces:
        - "application/json"
      responses:
        200:
          description: "The domain user with the specified name."
          schema:
            $ref: "#/definitions/DomainUser"
        400:
          description: "Bad request."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        404:
          description: "User not found."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        500:
          description: "Fatal server error."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
    patch:
      tags:
        - "domain-controller-connector"
      summary: "Updates a domain user."
      operationId: "updateUser"
      parameters:
        - name: "userName"
          description: "The user name of the domain user."
          type: "string"
          required: true
          in: "path"
        - name: "domainUser"
          description: "The update data of the domain user."
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/DomainUser"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        200:
          description: "The updated domain user."
          schema:
            $ref: "#/definitions/DomainUser"
        400:
          description: "Bad request."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        404:
          description: "User not found."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        500:
          description: "Fatal server error."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
    delete:
      tags:
        - "domain-controller-connector"
      summary: "Deletes a domain user."
      operationId: "deleteUser"
      parameters:
        - name: "userName"
          description: "The user name of the domain user."
          type: "string"
          required: true
          in: "path"
      produces:
        - "application/json"
      responses:
        200:
          description: "User does not exist or was successfully deleted."
        400:
          description: "Bad request."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        404:
          description: "User not found."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
        500:
          description: "Fatal server error."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
  /api/users/{userName}/f/exists:
    get:
      tags:
        - "domain-controller-connector"
      summary: "Checks whether a domain user exists."
      operationId: "userExists"
      parameters:
        - name: "userName"
          description: "The user name of the domain user."
          type: "string"
          required: true
          in: "path"
      produces:
        - "application/json"
      responses:
        200:
          description: "True if the user exists otherwise false."
          schema:
            $ref: "#/definitions/BooleanWrapper"
        500:
          description: "Fatal server error."
          schema:
            $ref: './META-INF/swagger/exception-model.yml#/definitions/RestApiException'
...
definitions:
  BooleanWrapper:
    description: "Wrapper of a boolean value."
    type: "object"
    properties:
      value:
        type: "boolean"
        description: "The boolean value."
    required:
      - value
...
  Names:
    type: "object"
    description: "A list with names."
    properties:
      values:
        type: "array"
        items:
          $ref: "#/definitions/Name"
        description: "The names."
...
  DomainGroupItem:
    description: "Domain group item."
    allOf:
      - $ref: "#/definitions/Common"
      - type: "object"
        properties:
          name:
            type: "string"
            description: "The name of the domain group."
    discriminator: "_type"
  DomainGroup:
    description: "Domain group."
    allOf:
      - $ref: "#/definitions/DomainGroupItem"
      - type: "object"
        properties:
          members:
            type: "array"
            items:
              $ref: "#/definitions/Name"
            description: "The members of the domain group."
  DomainUser:
    description: "Domain user."
    allOf:
      - $ref: "#/definitions/Common"
      - type: "object"
        properties:
          userName:
            type: "string"
            description: "The user name of the domain user."
          enabled:
            type: "boolean"
            description: "Is the domain user enabled?"
          passwordLastSet:
            type: "string"
            format: "date-time"
          displayName:
            type: "string"
          email:
            type: "string"
            format: "email"
            description: "The mail address of the domain user."
          mobile:
            type: "string"
            description: "The mobile number of the domain user."
          groups:
            type: "array"
            items:
              $ref: "#/definitions/Name"
            description: "The groups of the domain user."
  DomainUserAddRequest:
    description: "Domain user to add."
    allOf:
      - $ref: "#/definitions/DomainUser"
      - type: "object"
        properties:
          password:
            type: "string"
            format: "password"
            description: "The password of the domain user."
...